+++ poem-openapi/tests/api.rs
@@ -254,54 +254,6 @@ async fn payload_request() {
 }
 
 #[tokio::test]
-async fn optional_payload_request() {
-    struct Api;
-
-    #[OpenApi]
-    impl Api {
-        #[oai(path = "/", method = "post")]
-        async fn test(&self, req: Json<Option<i32>>) -> PlainText<String> {
-            PlainText(req.0.unwrap_or(999).to_string())
-        }
-    }
-
-    let meta: MetaApi = Api::meta().remove(0);
-    let meta_request = meta.paths[0].operations[0].request.as_ref().unwrap();
-    assert!(!meta_request.required);
-
-    assert_eq!(meta_request.content[0].content_type, "application/json");
-    assert_eq!(meta_request.content[0].schema, i32::schema_ref());
-
-    let ep = OpenApiService::new(Api, "test", "1.0").into_endpoint();
-    let resp = ep
-        .call(
-            poem::Request::builder()
-                .method(Method::POST)
-                .uri(Uri::from_static("/"))
-                .content_type("application/json")
-                .body("100"),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-    assert_eq!(resp.into_body().into_string().await.unwrap(), "100");
-
-    let ep = OpenApiService::new(Api, "test", "1.0").into_endpoint();
-    let resp = ep
-        .call(
-            poem::Request::builder()
-                .method(Method::POST)
-                .uri(Uri::from_static("/"))
-                .content_type("application/json")
-                .finish(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-    assert_eq!(resp.into_body().into_string().await.unwrap(), "999");
-}
-
-#[tokio::test]
 async fn response() {
     #[derive(ApiResponse)]
     enum MyResponse {