+++ poem/src/listener/acme/mod.rs
@@ -1,10 +1,12 @@
 //! Types for ACME.
 //!
 //! Reference: <https://datatracker.ietf.org/doc/html/rfc8555>
+//! Reference: <https://datatracker.ietf.org/doc/html/rfc8737>
 
 mod auto_cert;
 mod builder;
 mod client;
+mod endpoint;
 mod jose;
 mod keypair;
 mod listener;
@@ -15,6 +17,7 @@ mod serde;
 pub use auto_cert::AutoCert;
 pub use builder::AutoCertBuilder;
 pub use listener::{AutoCertAcceptor, AutoCertListener};
+pub use protocol::ChallengeType;
 
 /// Let's Encrypt production directory url
 pub const LETS_ENCRYPT_PRODUCTION: &str = "https://acme-v02.api.letsencrypt.org/directory";